Update 4:46 p.m.  Sample Application Using Universal Windows Platform c++/WinRT library for UI.
I have considered using this the C++/WinRT Library and creating a UWP project, since all of the UI components needed to build a good looking UI for an application that can be published to the Windows Store and run on any type of device that uses Windows 10 and have full access to UI Components defined with XAML and access to any hardware component via the WinRT library are included if we build an app on this platform.

There may also be a way to port this to Linux - I am not 100% sure yet.

This is a very simple project which contains an APP with 2 pages - "MainPage" and the "Start Page".  I have finally figured out how to navigate to the StartPage from the MainPage.  All components are specified in XAML, and handler methods are added to each of the runtime page classes which handle user interractions with the Controls (for instance, the user clicks on the "Start Tab", on "MainPage" and the handler for MainPage is invoked, etc.).

In order to build and run this project, you need to
1)Go to the Visual Studio installer and enable the Workloads "Universal Windows Platform Development" components.
2) Then go to the individual components and make sure to check "C++ Universal Windows Platform support for v142 tools" component to install, since C++ WinRt is not enabled by default with UWP.
3) Enable Developer Mode in the Secuity Settings of Windows 10, so that the application can be automatically installed after it is built.
4) You need to Start Visual Studio without creating a project - Just select "Run with no code"
5) Go to the Visual Studio Marketplace  via the Extensions Menu-> Manage Extensions,
and search for WinRT and install the C++/WinRT Tools.
6)  Hopefully, now, you should be able to build the project and run it within Visual Studio 2019 for Windows.

-You should be able to use all of the features of the C++ STL and WinRT and have complete access to a full API reference with myriad documentation on how to use each of the components of the WinRT API.

https://docs.microsoft.com/en-us/uwp/

I thought this was a better idea than trying to hack it out ourselves and running into a million different odd types of Windows errors and not having a common source of reference for interracting with Windows 10 OS.

If you think this is a bad idea, let me know.  It was a 2 day nightmare trying to figure out how to get this working, but once you figure it out, design should be pretty straight forward, and there are a million coding examples on how to use each of the various components and methods within the documentation; its compatible, and debugging is not impossible and you won't get screwy errors out of nowhere and try multiple solutions and not get anything to work.  Plus, you have access to OS components, like a FileOpenPicker dialog to allow you to select and open files on your system, without trying to go through the Common Windows Dialog with a thousand lines of cryptic old-fashioned C code which may not work correctly and requires you to do a great deal of hand-rolling your own code just to do simple operations, which is prone to errors, takes forever to do, and makes maintenance and debugging a nightmare if it doesn't work correctly - when WinRT already has a method defined which just needs to be called and has been thoroughly tested.  The only disadvantage is the project size is pretty large, 1.8 GB, but that is not such a big deal nowadays.