12:08 p.m. --Update.  Fixed openFile() bug and saveFile() bug.
If the user attempted to open a file and/or save a file after previously having a file open, and then decided to "Cancel" the save and/or open dialogs, the application would hang and/or not respond.  In addition, if the user tried to initially open a file, the previous bug fix had prevented the user from even being able to open a file.  Finally, the save_dialog() "Cancel" bug was fixed, but the open_dialog() on "Cancel" cleared the SourceCode file viewer, and if the user accidentally forgot to save their settings, the window was cleared and all unsaved changes were lost.  This was a bad scenario, and this is now fixed, after spending 12 hours trying to fix the change, and about 4 out of those 12 hours dealing with a corrupted source-file, that I wasn't aware of which never seemed to work correctly (the cursor was jumping all over the screen), until 15 minutes ago, I finally decided that particular file, itself, was the issue and not any of the changes I had made.  So, everything works now, until I can find something else that is wrong.